moving CompoundDocument to Utils.Storage namespace
